backend/requirements.txt
flask
flask-socketio
flask-cors
google-generativeai
python-dotenv
qrcode[pil]
